:recycle: Move unique key generation to top level trfm stage
